Refactored DatabaseCommands.
